+++ src/xcode/ENA/ENA/Source/Workers/Store/CacheData.swift
@@ -1,0 +1,11 @@
+//
+// 🦠 Corona-Warn-App
+//
+
+import Foundation
+
+struct CacheData: Codable {
+	let data: Data
+	let eTag: String
+	let date: Date
+}
